refactor: :recycle: ReviewDetail 페이지에서 사진유무로 스타일처리
